+++ api/src/school/domain/usecases/create-mission-assessment.js
@@ -1,4 +1,4 @@
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
+import { Assessment } from '../../../shared/domain/models/Assessment.js';
 
 const createMissionAssessment = async function ({ missionId, assessmentRepository }) {
   const assessment = Assessment.createForPix1dMission({ missionId });
