+++ api/tests/school/unit/application/assessments/assessment-controller_test.js
@@ -1,6 +1,7 @@
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 import { assessmentController } from '../../../../../src/school/application/assessments/assessment-controller.js';
 import { usecases } from '../../../../../src/school/shared/usecases/index.js';
+
 describe('Unit | Controller | assessment-controller', function () {
   describe('#getNextChallengeForPix1d', function () {
     it('should call the expected usecase', async function () {
@@ -25,6 +26,24 @@ describe('Unit | Controller | assessment-controller', function () {
 
       // then
       expect(result).to.be.equal(challenge);
+    });
+  });
+
+  describe('#createForPix1d', function () {
+    it('should call the expected usecase', async function () {
+      const missionId = 'mission-id';
+      const assessmentSerializer = { serialize: sinon.stub() };
+      const createdAssessment = Symbol('created-assessment');
+      assessmentSerializer.serialize.withArgs(createdAssessment).resolves(Symbol('serialized-assessment'));
+      sinon.stub(usecases, 'createMissionAssessment').withArgs({ missionId }).resolves(createdAssessment);
+      const request = { payload: { missionId } };
+
+      const result = await assessmentController.createForPix1d(request, hFake, {
+        assessmentSerializer,
+      });
+
+      expect(result.statusCode).to.be.equal(201);
+      expect(assessmentSerializer.serialize).to.have.been.calledWithExactly(createdAssessment);
     });
   });
 });
