+++ api/tests/unit/application/assessments/assessment-controller_test.js
@@ -7,23 +7,6 @@ import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransact
 import { Activity } from '../../../../src/school/domain/models/Activity.js';
 
 describe('Unit | Controller | assessment-controller', function () {
-  describe('#createForPix1d', function () {
-    it('should call the expected usecase', async function () {
-      const missionId = 'mission-id';
-      const assessmentSerializer = { serialize: sinon.stub() };
-      const createdAssessment = Symbol('created-assessment');
-      assessmentSerializer.serialize.withArgs(createdAssessment).resolves(Symbol('serialized-assessment'));
-      sinon.stub(usecases, 'createMissionAssessment').withArgs({ missionId }).resolves(createdAssessment);
-      const request = { payload: { missionId } };
-
-      const result = await assessmentController.createForPix1d(request, hFake, {
-        assessmentSerializer,
-      });
-
-      expect(result.statusCode).to.be.equal(201);
-      expect(assessmentSerializer.serialize).to.have.been.calledWithExactly(createdAssessment);
-    });
-  });
   describe('#createAssessmentPreviewForPix1d', function () {
     it('should call the expected usecase', async function () {
       const assessmentSerializer = { serialize: sinon.stub() };
